Move CairoOp to ast.rs
